requirements.txt
quart
grequests
beautifulsoup4
alive-progress
surrealdb
pytz
python-dotenv